float->double & SP_TIME define

+++ rl-learner/src/main.cpp
@@ -46,9 +46,9 @@ int main(int argc, char *argv[])
 			currentEpsilon = 0.0f;
 		}
 #ifdef USE_NEW_REWARD_CODE
-		int numTerm = 1;
+		unsigned int numTerm = 1;
 #else
-		int numTerm = RL_LAB_INDEX; //1
+		unsigned int numTerm = RL_LAB_INDEX; //1
 #endif
 		for ( unsigned int cTerm = 0 ; cTerm < numTerm ; cTerm++ )
 		{
@@ -371,7 +371,7 @@ void LoadConfig(int argc, char *argv[])
 		ALPHA = pt.get("alpha", ALPHA);
 		EPSILON_START = pt.get("epsilon_start", EPSILON_START);
 		EPSILON_DECAY = pt.get("epsilon_decay", EPSILON_DECAY);
-		float epsilon_end = pt.get("epsilon_end", 0.0);
+		double epsilon_end = pt.get("epsilon_end", 0.0);
 		if(epsilon_end != 0.0)
 		{
 			double first = (double)epsilon_end/(double)EPSILON_START;
@@ -395,9 +395,9 @@ void LoadConfig(int argc, char *argv[])
 			("num_learners,l", po::value<int>(), "Number of learners")
 			("gamma,g", po::value<float>(), "Gamma value")
 			("alpha,a", po::value<float>(), "Alpha value")
-			("epsilon_start,e", po::value<float>(), "Starting epsilon value")
-			("epsilon_decay,d", po::value<float>(), "Epsilon-decay value")
-			("epsilon_end,f", po::value<float>(), "Epsilon-end value(overwrites decay)")
+			("epsilon_start,e", po::value<double>(), "Starting epsilon value")
+			("epsilon_decay,d", po::value<double>(), "Epsilon-decay value")
+			("epsilon_end,f", po::value<double>(), "Epsilon-end value(overwrites decay)")
 			("print_reward,r", po::value<bool>(), "Print reward? Else time is pritned.")
 			("use_smdpq,s", po::value<bool>(), "Use smdpq?")
 			("use_rs_termination,t", po::value<bool>(), "Use special termination reward?")
@@ -426,13 +426,13 @@ void LoadConfig(int argc, char *argv[])
 			ALPHA = vm["alpha"].as<float>();
 		}
 		if (vm.count("epsilon_start")) {
-			EPSILON_START = vm["epsilon_start"].as<float>();
+			EPSILON_START = vm["epsilon_start"].as<double>();
 		}
 		if (vm.count("epsilon_decay")) {
-			EPSILON_DECAY = vm["epsilon_decay"].as<float>();
+			EPSILON_DECAY = vm["epsilon_decay"].as<double>();
 		}
 		if (vm.count("epsilon_end")) {
-			float epsilon_end = vm["epsilon_end"].as<float>();
+			double epsilon_end = vm["epsilon_end"].as<double>();
 			EPSILON_DECAY = pow((double)epsilon_end/(double)EPSILON_START, 1.0/(double)RUNS);
 		}
 		if (vm.count("print_reward")) {
